# apps/castaway-web/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: Castaway Web API
  version: 0.0.0
...
paths:
...
  /instances/{instanceID}/finale-bingo/loan-sharks:
    post:
      operationId: recordFinaleBingoLoanSharks
      parameters:
        - name: instanceID
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                anyOf:
                  - type: object
                    additionalProperties: {}
                  - $ref: '#/components/schemas/ErrorResponse'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecordFinaleBingoLoanSharksRequest'
  /instances/{instanceID}/finale-bingo/scores:
    post:
      operationId: recordFinaleBingoScores
      parameters:
        - name: instanceID
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                anyOf:
                  - type: object
                    additionalProperties: {}
                  - $ref: '#/components/schemas/ErrorResponse'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecordFinaleBingoScoresRequest'
  /instances/{instanceID}/finale-bingo/scores/preview:
    post:
      operationId: previewFinaleBingoScores
      parameters:
        - name: instanceID
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                anyOf:
                  - type: object
                    additionalProperties: {}
                  - $ref: '#/components/schemas/ErrorResponse'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecordFinaleBingoScoresRequest'
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
    FinaleBingoLoanSharkInput:
      type: object
      required:
        - shark_participant_id
        - target_participant_id
      properties:
        shark_participant_id:
          type: string
        target_participant_id:
          type: string
    FinaleBingoScoreInput:
      type: object
      required:
        - participant_id
        - box_points
        - bingo_count
      properties:
        participant_id:
          type: string
        box_points:
          type: integer
          format: int32
        bingo_count:
          type: integer
          format: int32
        notes:
          type: string
...
    RecordFinaleBingoLoanSharksRequest:
      type: object
      required:
        - assignments
      properties:
        name:
          type: string
        assignments:
          type: array
          items:
            $ref: '#/components/schemas/FinaleBingoLoanSharkInput'
        effective_at:
          type: string
          format: date-time
    RecordFinaleBingoScoresRequest:
      type: object
      required:
        - scores
      properties:
        name:
          type: string
        scores:
          type: array
          items:
            $ref: '#/components/schemas/FinaleBingoScoreInput'
        loan_sharks:
          type: array
          items:
            $ref: '#/components/schemas/FinaleBingoLoanSharkInput'
        effective_at:
          type: string
          format: date-time
...
servers:
  - url: http://localhost:8080
    description: Local development
    variables: {}